requirements.txt
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.13.0b2
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chromedriver==2.24.1
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-DebugToolbar==0.15.1
Flask-Mailman==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==1.0.0
GeoAlchemy2==0.15.2
google==3.0.0
google-api-core==2.19.2
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.0
grpcio-status==1.66.0
gunicorn==23.0.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httplib2==0.22.0
idna==3.8
inotify==0.2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
lxml==5.3.0
MarkupSafe==2.1.5
nose==1.3.7
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
probableparsing==0.0.1
proto-plus==1.24.0
protobuf==5.27.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pymongo==4.8.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySocks==1.7.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
xvfbwrapper==0.2.9
